week-08/requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.28.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0